--- frontend/packages/frontend/declarations/universal-cookie.d.ts
@@ -1,1 +1,0 @@
-declare module 'universal-cookie';
