+++ frontend/packages/frontend/src/admin/PackContext.tsx
@@ -1,10 +1,10 @@
 import React, {
   useState, useEffect, createContext, ReactNode,
 } from "react";
-import { client } from "src/TerryClient";
-import { Loadable } from "src/Loadable";
+import { client } from "@terry/shared/_/TerryClient";
+import { Loadable } from "@terry/shared/_/Loadable";
 import { useTriggerUpdate } from "@terry/shared/_/hooks/useTriggerUpdate";
-import { notifyError } from "src/utils";
+import { notifyError } from "@terry/shared/_/utils";
 import { Pack } from "./types";
 
 type PackContextType = {
