+++ frontend/packages/frontend/src/admin/hooks/useUsers.ts
@@ -1,9 +1,9 @@
 import { AxiosResponse } from "axios";
 import { useState, useEffect } from "react";
-import { Loadable } from "src/Loadable";
-import { client } from "src/TerryClient";
+import { Loadable } from "@terry/shared/_/Loadable";
+import { client } from "@terry/shared/_/TerryClient";
 import { useToken } from "src/admin/AdminContext";
-import { notifyError } from "src/utils";
+import { notifyError } from "@terry/shared/_/utils";
 import { useTriggerUpdate } from "@terry/shared/_/hooks/useTriggerUpdate";
 import { UsersData } from "src/admin/types";
