+++ frontend/packages/frontend/src/contest/hooks/useStatement.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect, useMemo } from "react";
-import { Loadable } from "src/Loadable";
-import { client } from "src/TerryClient";
+import { Loadable } from "@terry/shared/_/Loadable";
+import { client } from "@terry/shared/_/TerryClient";
 
 export function useStatement(path: string) {
   const [statement, setStatement] = useState<Loadable<string>>(Loadable.loading());
