+++ frontend/packages/frontend/src/contest/task/TaskView.tsx
@@ -7,7 +7,7 @@ import { CreateSubmissionView } from "src/contest/submission/submit/CreateSubmis
 import { SubmissionReportView } from "src/contest/submission/SubmissionReportView";
 import { SubmissionListView } from "src/contest/submission/SubmissionListView";
 import { useSubmissionList } from "src/contest/hooks/useSubmissionList";
-import { Loading } from "src/Loading";
+import { Loading } from "@terry/shared/_/components/Loading";
 import { TaskCommands } from "./TaskCommands";
 import { TaskStatementView } from "./TaskStatementView";
 import { LastSubmission } from "./LastSubmission";
