+++ frontend/packages/shared/declarations/universal-cookie.d.ts
@@ -1,0 +1,1 @@
+declare module "universal-cookie";
