+++ frontend/packages/shared/package.json
@@ -21,10 +21,16 @@
         "@lingui/react": "^2.9.2",
         "@types/react": "^16.7.18",
         "@types/react-dom": "^16.0.11",
+        "@types/react-modal": "^3.1.2",
+        "@types/react-router-dom": "^5.1.5",
         "luxon": "^1.0.0",
         "moment": "^2.22.1",
         "react": "^16.7.0",
         "react-dom": "^16.7.0",
+        "react-router-dom": "^5.1.5",
+        "axios": "^0.18.0",
+        "universal-cookie": "^2.1.2",
+        "react-modal": "^3.4.1",
         "typescript": "^3.2.2"
     },
     "devDependencies": {
