+++ frontend/packages/shared/src/utils.ts
