+++ frontend/packages/shared/tsconfig.json
@@ -23,7 +23,8 @@
     "allowJs": true
   },
   "include": [
-    "src/**/*"
+    "src",
+    "declarations"
   ],
   "exclude": [
     "node_modules",
